refactor(configstate): split monolithic file by responsibility
Break the single 590-line configstate.go into focused files that share one
package-identity rule, so a change to one concern (channel resolution, merge,
pin) no longer touches the others:

- identity.go: package-identity keys and name normalization
- channel.go:  saved-channel index and effective-channel resolution
- merge.go:    merge resolved packages back into gup.json, version persistence
- pin.go:      add/remove concrete version pins
- configstate.go: read/validate/resolve entry points the cmd/ layer calls

Pure code move; no behavior change. Existing tests pass unchanged.
